requirements.txt
pytube>=15.0.0
moviepy>=2.0.0
transformers>=4.0.0
torch>=2.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.25.0
beautifulsoup4>=4.10.0
yt-dlp>=2023.1.0
openai-whisper>=20230918